Explosions now (mostly) work properly when destroying stairs. They destroy both the up and down stairs (generating the next level if need be). If an agent is standing on the floor when it collapses, they fall to the next level.
Three things: the agent still falls into the explosion on the next floor
which I don't think I want. Need to test what happens if the player is
standing on the up stairs when the bomb goes off and a monsters falls
through from the above floor. (The only scenario I can see this happening
in is if the player parks a robot on the stairs above) I don't think I
am calculating which stairs to affect for the part of the dungeon where
higher floor #s are reached via the up stairs.